chore: add type check[]
